+++ lib-impl/src/test/java/org/ergoplatform/appkit/impl/BlockchainContextImplTest.java
@@ -174,7 +174,7 @@ public class BlockchainContextImplTest extends ApiTestBase {
         }
 
         @Override
-        public long getMaxBlockCost() {
+        public int getMaxBlockCost() {
             return 0;
         }
